REFACT - Adressed compiler warnings - organised import statements; - removed declared but unthrown exceptions; - tidied unnecessary else statements; - added `this` qualifers for class member references; - fixed failure to close `Scannner` in `FeatureTester`; - made methods static where possible; and - addressed boxing and unboxing warnings.

+++ validation-model/src/main/java/org/verapdf/gf/model/impl/pd/GFPDHalftone.java
@@ -102,7 +102,7 @@ public class GFPDHalftone extends GFPDObject implements org.verapdf.model.pdlaye
 
     private List<org.verapdf.model.pdlayer.PDHalftone> getHalftones() {
         Long halftoneType = getHalftoneType();
-        if (halftoneType == null || halftoneType != 5L) {
+        if (halftoneType == null || halftoneType.longValue() != 5L) {
             return Collections.emptyList();
         }
         List<org.verapdf.model.pdlayer.PDHalftone> halftones = new ArrayList<>();
